refactor(browse): move fetchPostsAndTags into findProfileStore
Extract the /browse/bounds API call from useBrowseViewModel into
findProfileStore so the composable no longer owns fetch logic,
abort controllers, or loading state. It now reads postPois,
availableTags, and isLoadingPosts from the store via storeToRefs.

Also extract bounds utility functions (boundsContain, padBounds,
unionBounds, profileInBounds) into browse/utils/boundsUtils.ts.
