+++ apps/frontend/src/features/browse/composables/useBrowseViewModel.ts
@@ -1,15 +1,13 @@
 import { computed, ref, type Ref } from 'vue'
-import { api, safeApiCall } from '@/lib/api'
-import type { BrowseBoundsResponse } from '@zod/apiResponse.dto'
-import type { PublicTag } from '@zod/tag/tag.dto'
+import { storeToRefs } from 'pinia'
 import type {
-  MapBounds,
   MapCluster,
   MapPoi,
   BoundsWithZoom,
 } from '@/features/map/types/map.types'
-import type { PublicPostWithProfile } from '@zod/post/post.dto'
 import type { ClusterFeature, MapFeature, PointFeature } from '@shared/zod/map/cluster.dto'
+import { useFindProfileStore } from '@/features/browse/stores/findProfileStore'
+
 /**
  * Composable that manages the posts data layer, bounds-scoped tags,
  * and map selection state for the unified browse map. Profile clustering
@@ -19,53 +17,8 @@ export function useBrowseViewModel(
   clusterFeatures: Ref<MapFeature[]>,
   onProfileBoundsChanged: (b: BoundsWithZoom) => void
 ) {
-  const postPois = ref<MapPoi[]>([])
-  const availableTags = ref<PublicTag[]>([])
-  const isLoadingPosts = ref(false)
-  let postAbortController: AbortController | null = null
-
-  async function fetchPostsAndTags(bounds: MapBounds) {
-    if (postAbortController) postAbortController.abort()
-    const controller = new AbortController()
-    postAbortController = controller
-
-    // TODO move this into pinia store
-    isLoadingPosts.value = true
-    try {
-      const res = await safeApiCall(() =>
-        api.get<BrowseBoundsResponse>('/browse/bounds', {
-          params: bounds,
-          signal: controller.signal,
-        })
-      )
-
-      if (!res.data.success) return
-
-      // Map posts to MapPoi[]
-      postPois.value = (res.data.posts as PublicPostWithProfile[])
-        .filter((p) => p.location?.lat != null && p.location?.lon != null)
-        .map((p) => ({
-          id: p.id,
-          title: p.content?.substring(0, 50) ?? '',
-          location: { lat: p.location!.lat!, lon: p.location!.lon! },
-          image: p.postedBy?.profileImages?.[0],
-          type: 'post',
-          source: p,
-        }))
-
-      // Update available tags for the filter pill
-      availableTags.value = res.data.tags
-    } catch (err: any) {
-      if (err?.code === 'ERR_CANCELED') return
-      // On transient errors, keep the last good postPois/availableTags on
-      // screen rather than flashing to empty — the user just panned and a
-      // blank map is worse UX than slightly-stale results.
-    } finally {
-      if (postAbortController === controller) {
-        isLoadingPosts.value = false
-      }
-    }
-  }
+  const findProfileStore = useFindProfileStore()
+  const { postPois, availableTags, isLoadingPosts } = storeToRefs(findProfileStore)
 
   const clusters = computed<MapCluster[]>(() =>
     clusterFeatures.value
@@ -78,6 +31,7 @@ export function useBrowseViewModel(
       }))
   )
 
+  // TODO type wrangling - what type are we converting into?
   // this belongs to the pinia store, or better yet, into the backend
   const profilePois = computed<MapPoi[]>(() =>
     clusterFeatures.value
@@ -108,7 +62,7 @@ export function useBrowseViewModel(
   async function onBoundsChanged(boundsWithZoom: BoundsWithZoom) {
     await Promise.all([
       onProfileBoundsChanged(boundsWithZoom),
-      fetchPostsAndTags(boundsWithZoom.bounds),
+      findProfileStore.fetchPostsAndTags(boundsWithZoom.bounds),
     ])
   }
 
@@ -119,7 +73,6 @@ export function useBrowseViewModel(
     allPois,
     availableTags,
     isLoadingPosts,
-    fetchPostsAndTags,
     activePoi,
     onSelectionClear,
     onBoundsChanged,
